requirements.txt
pyserial>=3.5
RPi.GPIO>=0.7.0
